+++ src/tests_integration/mod.rs
@@ -1,4 +1,5 @@
 mod catchers;
+mod db_notification;
 mod db_online;
 mod errors_api_error;
 mod keepalive;